services/coral-app/requirements/common.txt
aiohttp==3.7.4
tornado==6.1
ffmpeg-python==0.2.0
imageio==2.9.0
opencv-contrib-python==4.5.3.56
pyzmq==22.2.1
asyncio==3.4.3
numpy==1.21.2